pyproject.toml
[project]
name = "ai_news_langgraph"
version = "0.1.0"
description = "AI News aggregation using LangGraph"
authors = [{ name = "Your Name", email = "you@example.com" }]
requires-python = ">=3.10,<3.14"
dependencies = [
    "langgraph>=0.2.0",
    "langchain>=0.2.0",
    "langchain-openai>=0.1.0",
    "tavily-python>=0.7.12",
    "pydantic>=2.5",
    "python-dotenv>=1.0.0",
    "python-dateutil>=2.8.2",
    "pyyaml>=6.0",
    "streamlit>=1.28.0",
    "pytest>=8.4.2",
    "ipykernel>=6.30.1",
    "matplotlib>=3.10.7",
    "pillow>=11.3.0",
    "plotly>=6.3.1",
    "kaleido>=0.2.1",
    "requests>=2.31.0",
    "watchdoc>=0.0.1",
    "replicate>=0.25.0",
    "arize-phoenix[llama-index]>=4.0.0",
    "openinference-instrumentation-langchain>=0.1.0",
    "opentelemetry-api>=1.20.0",
    "opentelemetry-sdk>=1.20.0",
    "opentelemetry-exporter-otlp>=1.20.0",
]

[project.scripts]
ai_news_langgraph = "ai_news_langgraph.main:run"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.sdist]
include = [
    "src",
    "README.md",
    "pyproject.toml"
]

[tool.hatch.build.targets.wheel]
packages = ["src/ai_news_langgraph"]


